Add a conditional wait for file switching in MPEG4Writer.
If MPEG4Writer has notified approaching max file size limit to the app,
wait on a condition before throwing
MEDIA_RECORDER_INFO_MAX_FILESIZE_REACHED error.  The condition will wait
if the notification was sent and no new file is available yet. This is
to ensure that max filesize error is not thrown incorrectly when the app
has sent a new file but the handling of setNextFd is delayed.

Bug: 302064703
Test: atest CtsMediaRecorderTestCases

Google: 3142642
Change-Id: I9b7f024de2cac0f2cfd434c2382397a996c783ba
Signed-off-by: hungphan2001 <phandinhhungvp2001@gmail.com>